Update existing BigQuery sample
