+++ bigquery/Gemfile
@@ -1,4 +1,4 @@
-# Copyright 2015 Google, Inc
+# Copyright 2016 Google, Inc
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,10 +14,11 @@
 
 source "https://rubygems.org"
 
-gem "gcloud"
+gem "google-cloud-bigquery"
 
 group :test do
   gem "rake"
   gem "rubocop"
   gem "rspec"
+  gem "google-cloud-storage"
 end
