+++ bigquery/spec/bigquery_sample_spec.rb
