--- bigquery/spec/data/test_data.csv
@@ -1,2 +1,0 @@
-"Alice","Human"
-"Caterpillar", "Butterfly Larva"
