--- bigquery/spec/data/test_data.json
@@ -1,2 +1,0 @@
-{ "name": "Alice", "title": "Human"  }
-{ "name": "Caterpillar", "title": "Butterfly Larva" }
