--- bigquery/spec/list_tables_spec.rb
@@ -1,23 +1,0 @@
-# Copyright 2015 Google, Inc
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-require_relative "spec_helper"
-require_relative "../list_tables"
-
-RSpec.describe "List tables sample" do
-  it "lists the dataset's tables" do
-    expect { list_tables PROJECT_ID, "test_dataset"}.to(
-      output(/test_table/).to_stdout)
-  end
-end
