--- bigquery_sample/.rspec
@@ -1,1 +1,0 @@
---tty --color --format documentation --order random
